+++ janitor-api/src/main/java/com/eischet/janitor/api/calls/JVoidMethod.java
@@ -11,5 +11,3 @@ import com.eischet.janitor.api.errors.runtime.JanitorRuntimeException;
 public interface JVoidMethod<T> {
     void call(final T self, final JanitorScriptProcess runningScript, final JCallArgs arguments) throws JanitorRuntimeException;
 }
-
-// TODO: there's already code for the same purpose in JBoundMethod/JUnboundMethod, so this is probably redundant. Consolidate!